requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.0